pyproject.toml
[project]
name = "retrieval-graph"
version = "0.0.1"
description = "Starter template for making a custom retrieval graph in LangGraph."
authors = [
    { name = "William Fu-Hinthorn", email = "13333726+hinthornw@users.noreply.github.com" },
]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph>=0.2.6",
    "langchain==0.3.23",
    "langchain-core==0.3.51",
    "langchain-openai==0.2.7",
    "langchain-elasticsearch==0.3.2",
    "langchain-ollama==0.2.3",
    "python-dotenv>=1.0.1",
    "msgspec>=0.18.6"
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["langgraph.templates.retrieval_graph", "retrieval_graph"]
[tool.setuptools.package-dir]
"langgraph.templates.retrieval_graph" = "src/retrieval_graph"
"retrieval_graph" = "src/retrieval_graph"

[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E", "F", "I", "D", "D401", "T201", "UP",
]
lint.ignore = [
    "UP006", "UP007", "UP035", "D417", "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"
